Extract ScoreType enum class
