+++ src/main/java/seedu/address/model/record/ScoreType.java
@@ -1,0 +1,8 @@
+package seedu.address.model.record;
+
+/**
+ * Represents the type of score being displayed.
+ */
+public enum ScoreType {
+    ATTENDANCE, SUBMISSION, PARTICIPATION
+}
